requirements.txt
pillow~=10.2.0
setuptools~=65.5.1